# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Micro Hackathon API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
    "/challenge/random": {
      "get": {
        "summary": "Get a random challenge",
        "parameters": [
          {
            "in": "query",
            "name": "difficulty",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Random challenge with ratings"
          }
        }
      }
    },
    "/challenge/of-the-day": {
      "get": {
        "summary": "Get the challenge of the day",
        "responses": {
          "200": {
            "description": "Featured challenge of the day"
          }
        }
      }
    },
    "/challenge/top": {
      "get": {
        "summary": "Get top-rated challenges",
        "responses": {
          "200": {
            "description": "Top challenges based on rating"
          }
        }
      }
    },
    "/challenge/tags": {
      "get": {
        "summary": "Get all challenge tags",
        "responses": {
          "200": {
            "description": "List of tags"
          }
        }
      }
    },
    "/challenge/all": {
      "get": {
        "summary": "Get all challenges",
        "responses": {
          "200": {
            "description": "List of all challenges"
          }
        }
      }
    },
    "/challenge/{id}/ratings": {
      "get": {
        "summary": "Get rating summary for a challenge",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Average rating and count"
          }
        }
      }
    },
    "/challenge/{id}/rate": {
      "post": {
        "summary": "Submit a rating",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "number"
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rating submitted"
          }
        }
      }
    },
    "/challenge": {
      "get": {
        "summary": "Get challenges by tag",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of challenges filtered by tag"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}